# openapi.yaml
openapi: 3.0.1
info:
  title: Freedom API
  description: Freedom Project API Documentation
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
  /admin/resource/update/resource/{resourceId}:
    put:
      tags:
      - resource-admin-controller
      operationId: updateResource
      parameters:
      - name: resourceId
        in: path
        required: true
        schema:
          type: string
      - name: title
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/update/material/{materialId}:
    put:
      tags:
      - resource-admin-controller
      operationId: updateMaterial
      parameters:
      - name: materialId
        in: path
        required: true
        schema:
          type: string
      - name: title
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/upload/resource:
    post:
      tags:
      - resource-admin-controller
      operationId: uploadResource
      parameters:
      - name: resourceId
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/upload/material:
    post:
      tags:
      - resource-admin-controller
      operationId: uploadMaterial
      parameters:
      - name: resourceId
        in: query
        required: true
        schema:
          type: string
      - name: materialId
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/addLabel/{resourceId}:
    post:
      tags:
      - resource-admin-controller
      operationId: addLabelToResource
      parameters:
      - name: resourceId
        in: path
        required: true
        schema:
          type: string
      - name: labelId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/label/linkChildren/{parentLabelId}:
    post:
      tags:
      - label-admin-controller
      operationId: linkChildrenLabels
      parameters:
      - name: parentLabelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int64
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/label/createOrUpdate:
    post:
      tags:
      - label-admin-controller
      operationId: updateLabel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Label'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /public/resource/download/{type}/{id}:
    get:
      tags:
      - resource-public-controller
      operationId: downloadFile
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /public/resource/all:
    get:
      tags:
      - resource-public-controller
      operationId: getAllResourceInfo
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /private/account/info:
    get:
      tags:
      - account-private-controller
      operationId: getAccountInfo
      responses:
        "200":
          description: OK
  /admin/resource/materials/{resourceId}:
    get:
      tags:
      - resource-admin-controller
      operationId: getAllMaterialsByResource
      parameters:
      - name: resourceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/labels/{resourceId}:
    get:
      tags:
      - resource-admin-controller
      operationId: getAllLabelsByResource
      parameters:
      - name: resourceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/all:
    get:
      tags:
      - resource-admin-controller
      operationId: getAllResources
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/label/nextLayers:
    get:
      tags:
      - label-admin-controller
      operationId: getNextLayerLabels
      parameters:
      - name: labelKey
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/label/all:
    get:
      tags:
      - label-admin-controller
      operationId: getAllLabels
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/removeLabel/{resourceId}:
    delete:
      tags:
      - resource-admin-controller
      operationId: removeLabelFromResource
      parameters:
      - name: resourceId
        in: path
        required: true
        schema:
          type: string
      - name: labelId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/deleteAllFile:
    delete:
      tags:
      - resource-admin-controller
      operationId: deleteAllFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/delete/resource/{resourceId}:
    delete:
      tags:
      - resource-admin-controller
      operationId: deleteResource
      parameters:
      - name: resourceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/resource/delete/material/{materialId}:
    delete:
      tags:
      - resource-admin-controller
      operationId: deleteMaterial
      parameters:
      - name: materialId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/label/unlinkChildren/{parentLabelId}/{childLabelId}:
    delete:
      tags:
      - label-admin-controller
      operationId: unlinkChildrenLabels
      parameters:
      - name: parentLabelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: childLabelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
  /admin/label/delete/{labelId}:
    delete:
      tags:
      - label-admin-controller
      operationId: deleteLabel
      parameters:
      - name: labelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOutput'
components:
  schemas:
    ApiResponseOutput:
      required:
      - code
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
    Label:
      required:
      - level
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        level:
          type: integer
          format: int32
        createdDate:
          type: string
          format: date-time
        updatedDate:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT